+++ src/Compute/Compute.Test/ScenarioTests/GalleryTests.cs
@@ -100,5 +100,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             TestRunner.RunTestScript("Test-GalleryImageDefinitionUpdateFeature");
         }
 
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGalleryManagedIdentity()
+        {
+            TestRunner.RunTestScript("Test-GalleryManagedIdentity");
+        }
+
     }
 }